fix(libsinsp): Add a CPU count check
In a scap file I have here, minfo->num_cpus is 0 when
sinsp_filter_check_thread::extract_exectime is called. We handle this in
debug builds with an ASSERT, but need to check for this in release
builds as well. Bail out early if we either have no CPUs or if minfo is
NULL.

Signed-off-by: Gerald Combs <gerald@wireshark.org>